@@ -1,7 +1,7 @@
 # https://github.com/rspec/rspec-rails/issues/255#issuecomment-7858480
 class ActionDispatch::Routing::RouteSet
-  def url_for_with_locale_fix(options)
-    url_for_without_locale_fix({ locale: I18n.locale }.merge(options))
+  def url_for_with_locale_fix(options, route_name = nil, url_strategy = UNKNOWN)
+    url_for_without_locale_fix({ locale: I18n.locale }.merge(options), route_name, url_strategy)
   end
 
   alias_method_chain :url_for, :locale_fix
@@ -9,7 +9,13 @@ end
 
 # https://github.com/rspec/rspec-rails/issues/255#issuecomment-20727452
 class ActionDispatch::Routing::RouteSet::NamedRouteCollection::UrlHelper
-  def call(t, args)
-    t.url_for(handle_positional_args(t, args, { locale: I18n.locale }.merge(@options), @segment_keys))
+  def call_with_locale_fix(t, args, inner_options)
+    original_options = @options
+    @options = { locale: I18n.locale }.merge(@options)
+    result = call_without_locale_fix(t, args, inner_options)
+    @options = original_options
+    result
   end
+
+  alias_method_chain :call, :locale_fix
 end
